refactor(translatable): views uses translatable provider
KK-1017.

Instead of using the old `useLanguageTabs` implementation, the views are now
using the new TranslatableProvider.

Also merged the venue edit and creation forms to a new component.
