+++ src/domain/eventGroups/edit/EventGroupsEdit.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 
 import KukkuuEditPage from '../../application/layout/kukkuuEditPage/KukkuuEditPage';
 import EventGroupForm from '../form/EventGroupForm';
+import TranslatableProvider from '../../../common/providers/TranslatableProvider';
 
 const EventGroupsEdit = () => {
   return (
@@ -9,7 +10,9 @@ const EventGroupsEdit = () => {
       pageTitleSource="name"
       reactAdminProps={{ redirect: '/events-and-event-groups' }}
     >
-      <EventGroupForm />
+      <TranslatableProvider>
+        <EventGroupForm />
+      </TranslatableProvider>
     </KukkuuEditPage>
   );
 };
