+++ src/domain/events/create/EventCreate.tsx
@@ -5,6 +5,7 @@ import { useLocation } from 'react-router-dom';
 import Aside from '../../../common/components/aside/Aside';
 import KukkuuCreatePage from '../../application/layout/kukkuuCreatePage/KukkuuCreatePage';
 import EventForm from '../eventForm/EventForm';
+import TranslatableProvider from '../../../common/providers/TranslatableProvider';
 
 const EventCreate = () => {
   const { search } = useLocation();
@@ -30,7 +31,9 @@ const EventCreate = () => {
         redirect,
       }}
     >
-      <EventForm view="create" />
+      <TranslatableProvider>
+        <EventForm view="create" />
+      </TranslatableProvider>
     </KukkuuCreatePage>
   );
 };
