+++ src/domain/messages/create/MessagesCreate.tsx
@@ -6,6 +6,7 @@ import { ProtocolType } from '../../../api/generatedTypes/globalTypes';
 import KukkuuCreatePage from '../../application/layout/kukkuuCreatePage/KukkuuCreatePage';
 import KukkuuCreateToolbar from '../../application/layout/kukkuuCreatePage/KukkuuCreateToolbar';
 import MessageForm from '../form/MessageForm';
+import TranslatableProvider from '../../../common/providers/TranslatableProvider';
 
 const MessageCreate = () => {
   const t = useTranslate();
@@ -28,16 +29,18 @@ const MessageCreate = () => {
         redirect: 'list',
       }}
     >
-      <MessageForm
-        toolbar={
-          <KukkuuCreateToolbar
-            saveLabel={
-              protocol === ProtocolType.SMS ? t('action.send') : undefined
-            }
-          />
-        }
-        protocol={protocol}
-      />
+      <TranslatableProvider>
+        <MessageForm
+          toolbar={
+            <KukkuuCreateToolbar
+              saveLabel={
+                protocol === ProtocolType.SMS ? t('action.send') : undefined
+              }
+            />
+          }
+          protocol={protocol}
+        />
+      </TranslatableProvider>
     </KukkuuCreatePage>
   );
 };
