+++ src/domain/messages/edit/MessagesEdit.tsx
@@ -2,11 +2,14 @@ import React from 'react';
 
 import KukkuuEditPage from '../../application/layout/kukkuuEditPage/KukkuuEditPage';
 import MessageForm from '../form/MessageForm';
+import TranslatableProvider from '../../../common/providers/TranslatableProvider';
 
 const MessagesEdit = () => {
   return (
     <KukkuuEditPage pageTitleSource="subject">
-      <MessageForm />
+      <TranslatableProvider>
+        <MessageForm />
+      </TranslatableProvider>
     </KukkuuEditPage>
   );
 };
